// demo-mcp-21-code-generator-node/input/reference-repos/MSPrChValGenerate/src/common/utils/enums/collectionsNames.enum.ts
/**
 * Enumeraciones de las colecciones del microservicio
 * @author Deivid Santiago Vargas
 */

export enum CollectionsNames {
  TERMINALES_LIBRES = "COLPRTFREEEQUIPPRICES",
  TERMINALES_KIT_PREPAGO = "COLPRTKITPREPAIDPRICES",
  TECNOLOGIA = "COLPRTTECHNOLOGYPRICES",
  CARACTERISTICAS = "COLPRTTTATTRIBUTES",
  HOGARESATTRIBUTES = "COLPRTHOMEATTRIBUTES", 
  HOGARESPRICES = "COLPRTHOMEPRICES",
  MOVILATTRIBUTES = "COLPRTPREPOATTRIBUTES", 
  MOVILPRICES = "COLPRTPREPOPRICES",
  EXCEPTION_SKU = "COLPRTEXCEPTIONS",
}